# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/job-openings/{jobOpeningId}/refferals/{referralId}": {
      "put": {
        "tags": [
          "job-opening-referral-controller"
        ],
        "operationId": "updateReferral",
        "parameters": [
          {
            "name": "jobOpeningId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "referralId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobOpeningReferralRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/job-openings/{id}": {
      "get": {
        "tags": [
          "job-opening-controller"
        ],
        "operationId": "get_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultJobOpeningResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "job-opening-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobOpeningRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/engagement/tags/{id}": {
      "get": {
        "tags": [
          "tag-controller"
        ],
        "operationId": "get_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultTagResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tag-controller"
        ],
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tag-controller"
        ],
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/job-openings": {
      "get": {
        "tags": [
          "job-opening-controller"
        ],
        "operationId": "getAll_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListJobOpeningSummaryResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "job-opening-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobOpeningRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/job-openings/{id}/share": {
      "post": {
        "tags": [
          "job-opening-controller"
        ],
        "operationId": "share",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareJobOpeningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/job-openings/referrals": {
      "post": {
        "tags": [
          "job-opening-referral-controller"
        ],
        "operationId": "createReferral",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobOpeningReferralRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/engagement/tags": {
      "get": {
        "tags": [
          "tag-controller"
        ],
        "operationId": "getAll_4",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListTagResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tag-controller"
        ],
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/job-openings/{id}/close": {
      "patch": {
        "tags": [
          "job-opening-controller"
        ],
        "operationId": "close",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {

          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateJobOpeningReferralRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "jobOpeningId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "cvUrl": {
            "type": "string",
            "minLength": 1
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "cvUrl",
          "email",
          "id",
          "jobOpeningId",
          "name"
        ]
      },
      "UpdateJobOpeningRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "designationId": {
            "type": "string",
            "format": "uuid"
          },
          "requiredExperience": {
            "type": "number",
            "format": "float",
            "minimum": 0
          },
          "jd": {
            "type": "string",
            "format": "binary"
          },
          "defaultHrId": {
            "type": "string",
            "format": "uuid"
          },
          "hrs": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          },
          "reviewers": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "defaultHrId",
          "designationId",
          "id"
        ]
      },
...
      "UpdateTagRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
...
      "CreateJobOpeningRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "designationId": {
            "type": "string",
            "format": "uuid"
          },
          "requiredExperience": {
            "type": "number",
            "format": "float",
            "minimum": 0
          },
          "jd": {
            "type": "string",
            "format": "binary"
          },
          "defaultHrId": {
            "type": "string",
            "format": "uuid"
          },
          "hrs": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          },
          "reviewers": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "defaultHrId",
          "designationId"
        ]
      },
      "ShareJobOpeningRequest": {
        "type": "object",
        "properties": {
          "jobOpeningId": {
            "type": "string",
            "format": "uuid"
          },
          "shareToEmail": {
            "type": "string",
            "format": "email",
            "minLength": 1
          }
        },
        "required": [
          "shareToEmail"
        ]
      },
      "CreateJobOpeningReferralRequest": {
        "type": "object",
        "properties": {
          "jobOpeningId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "cvUrl": {
            "type": "string",
            "minLength": 1
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "cvUrl",
          "email",
          "jobOpeningId",
          "name"
        ]
      },
...
      "CreateTagRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
...
      "DesignationSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "GlobalUserResponseSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfileResponseSummary"
          }
        }
      },
...
      "UserProfileResponseSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          }
        }
      },
...
      "JobOpeningSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "designation": {
            "$ref": "#/components/schemas/DesignationSummaryResponse"
          },
          "requiredExperience": {
            "type": "number",
            "format": "float"
          },
          "jdUrl": {
            "type": "string"
          },
          "defaultHr": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "createdBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "updatedBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "closed": {
            "type": "boolean"
          }
        }
      },
      "ResultListJobOpeningSummaryResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobOpeningSummaryResponse"
            }
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "JobOpeningReferralResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "referredBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "cvUrl": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "IN_REVIEW",
              "ACCEPTED",
              "REJECTED"
            ]
          }
        }
      },
      "JobOpeningResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "designation": {
            "$ref": "#/components/schemas/DesignationSummaryResponse"
          },
          "requiredExperience": {
            "type": "number",
            "format": "float"
          },
          "jdUrl": {
            "type": "string"
          },
          "defaultHr": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "hrs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GlobalUserResponseSummary"
            }
          },
          "reviewers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GlobalUserResponseSummary"
            }
          },
          "referrals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobOpeningReferralResponse"
            }
          },
          "createdBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "updatedBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "shareAudits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobOpeningShareAuditResponse"
            }
          },
          "closed": {
            "type": "boolean"
          }
        }
      },
      "JobOpeningShareAuditResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sharedBy": {
            "$ref": "#/components/schemas/GlobalUserResponseSummary"
          },
          "sharedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sharedToEmail": {
            "type": "string"
          }
        }
      },
      "ResultJobOpeningResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/JobOpeningResponse"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ResultListTagResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponse"
            }
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "TagResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultTagResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/TagResponse"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}